frames_worker/requirements.txt
absl-py==2.1.0
attrs==25.1.0
cffi==1.17.1
contourpy==1.3.1
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.56.0
jax==0.5.0
jaxlib==0.5.0
kiwisolver==1.4.8
matplotlib==3.10.0
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
frame-ble
frame-msg
bleak
numpy
pillow
